app/agents/meta_agents/optimization_agent.py
from typing import Any, Dict, List

import numpy as np
from sklearn.ensemble import RandomForestRegressor
from app.agents.meta_agents.meta_agents import MetaAgent
from app.tools.reasoning_engine import ReasoningEngine
from app.tools.ontology_reasoner import OntologyReasoner
from app.models.knowledge.knowledge_base import KnowledgeBase
from app.tools.llm_manager import LLMManager
from app.models.knowledge.ontology.ontology import Ontology
import asyncio

class OptimizationAgent(MetaAgent):
    """
    Optimizes the deployed MAS based on feedback and performance metrics.
    """
    def __init__(self, agent_id: str, name: str, api_key: str, llm_service: Any, agent_communication_service: Any):
        super().__init__(
            agent_id=agent_id,
            name=name,
            api_key=api_key,
            llm_service=llm_service,
            agent_communication_service=agent_communication_service
        )
        self.agent_type = "optimization"
        self.knowledge_base = KnowledgeBase()
        self.reasoning_engine = ReasoningEngine(self.knowledge_base, api_key)
        self.llm_manager = LLMManager()
        self.ontology = Ontology()  # Initialize with your optimization ontology
        self.ontology_reasoner = OntologyReasoner(self.llm_manager, self.ontology)
        self.optimization_model = RandomForestRegressor()
        self.historical_data = []

        self._init_beliefs()
        self._init_desires()
        self._init_goals()
        self._init_plans()

    def _init_beliefs(self):
        self.add_belief("Continuous optimization is crucial for MAS performance")
        self.add_belief("Historical data provides valuable insights for optimization")

    def _init_desires(self):
        self.add_desire("Improve overall MAS efficiency", priority=10)
        self.add_desire("Identify and resolve performance bottlenecks", priority=9)

    def _init_goals(self):
        self.add_goal("Analyze current MAS performance", priority=9)
        self.add_goal("Generate optimization suggestions", priority=8)
        self.add_goal("Implement and test optimizations", priority=8)

    def _init_plans(self):
        self.create_plan(
            self.goals[0].id,
            [
                "Collect performance metrics",
                "Identify bottlenecks",
                "Generate optimization strategies",
                "Implement optimizations",
                "Evaluate optimization impact"
            ]
        )

    async def reason(self):
        print("Starting reasoning process for optimization")
        
        current_beliefs = [belief.to_dict() for belief in self.beliefs]
        updated_beliefs = await self.reasoning_engine.reason({"beliefs": current_beliefs})
        
        for belief in updated_beliefs.get("beliefs", []):
            self.add_belief(belief["content"])

        new_knowledge = await self.ontology_reasoner.infer_new_knowledge()
        
        for concept in new_knowledge.get("new_concepts", []):
            self.add_belief(f"New optimization concept: {concept['name']} - {concept['description']}")
        
        for relationship in new_knowledge.get("new_relationships", []):
            self.add_belief(f"New optimization relationship: {relationship['name']} between {relationship['domain']} and {relationship['range']}")

        optimization_query = "What are the key areas for optimization in this MAS?"
        optimization_areas = await self.ontology_reasoner.answer_query(optimization_query)
        self.add_belief(f"Key optimization areas: {optimization_areas}")

        current_beliefs = [belief.to_dict() for belief in self.beliefs]
        new_desires = await self.reasoning_engine.generate_desires(current_beliefs)
        
        for desire in new_desires:
            self.add_desire(desire["description"], desire["priority"])

        print("Reasoning process for optimization completed")

    async def plan(self):
        print("Starting planning process for optimization")
        for goal in self.goals:
            if goal.description == "Generate optimization suggestions":
                self.create_plan(
                    goal.id,
                    [
                        "Analyze historical performance data",
                        "Identify performance patterns",
                        "Generate optimization hypotheses",
                        "Prioritize optimization strategies"
                    ]
                )
        
        current_state = self.get_current_state()
        optimized_plan = await self.reasoning_engine.reason_and_plan(self.goals[0].description, current_state)
        
        if optimized_plan.get("plan"):
            self.update_plan(self.goals[0].id, optimized_plan["plan"].steps)
        
        print("Planning process for optimization completed")

    async def execute(self):
        print("Starting execution process for optimization")
        for plan in self.plans:
            for task in plan.steps:
                if task.status == "pending":
                    print(f"Executing task: {task.description}")
                    try:
                        execution_result = await self.reasoning_engine.simulate_action(task.description, self.get_current_state())
                        self.update_task_status(task.id, "completed")
                        
                        for key, value in execution_result.items():
                            self.add_belief(f"Optimization task result - {key}: {value}")
                    except Exception as e:
                        print(f"Error executing optimization task {task.description}: {str(e)}")
                        self.update_task_status(task.id, "failed")
        print("Execution process for optimization completed")

    def get_current_state(self) -> Dict[str, Any]:
        return {
            "beliefs": [belief.to_dict() for belief in self.beliefs],
            "desires": [desire.to_dict() for desire in self.desires],
            "goals": [goal.to_dict() for goal in self.goals],
            "plans": [plan.to_dict() for plan in self.plans],
            "historical_data": self.historical_data
        }

    async def run(self):
        while True:
            await self.reason()
            await self.plan()
            await self.execute()
            await asyncio.sleep(60)  # Adjust the sleep time based on desired optimization frequency

    async def suggest_optimizations(self, bottlenecks: List[str]) -> Dict[str, Any]:
        """
        Suggest optimizations to address the identified bottlenecks in the MAS.

        Args:
            bottlenecks (List[str]): A list of identified bottlenecks in the MAS.

        Returns:
            Dict[str, Any]: A dictionary containing the suggested optimizations for each bottleneck.
        """
        optimizations = {}
        for bottleneck in bottlenecks:
            optimization = await self._generate_optimization(bottleneck)
            optimizations[bottleneck] = optimization
        return optimizations

    async def _generate_optimization(self, bottleneck: str) -> Dict[str, Any]:
        if "execution time" in bottleneck.lower():
            return {"action": "Increase concurrency", "details": "Implement parallel processing for the agent's tasks"}
        elif "memory usage" in bottleneck.lower():
            return {"action": "Optimize memory", "details": "Implement memory-efficient data structures or use lazy loading"}
        elif "cpu usage" in bottleneck.lower():
            return {"action": "Distribute workload", "details": "Distribute computationally intensive tasks across multiple agents"}
        elif "task completion rate" in bottleneck.lower():
            return {"action": "Improve task prioritization", "details": "Implement a more sophisticated task prioritization algorithm"}
        elif "communication latency" in bottleneck.lower():
            return {"action": "Optimize communication", "details": "Implement more efficient communication protocols or reduce message size"}
        elif "error rate" in bottleneck.lower():
            return {"action": "Enhance error handling", "details": "Implement more robust error handling and recovery mechanisms"}
        else:
            return {"action": "Further investigation", "details": "Conduct a detailed analysis of the bottleneck"}

    async def implement_optimizations(self, mas_implementation: Dict[str, Any], optimizations: Dict[str, Any]) -> Dict[str, Any]:
        """
        Implement the suggested optimizations in the MAS.

        Args:
            mas_implementation (Dict[str, Any]): The current implementation of the MAS.
            optimizations (Dict[str, Any]): The suggested optimizations to be implemented.

        Returns:
            Dict[str, Any]: The updated MAS implementation with the optimizations applied.
        """
        updated_mas = mas_implementation.copy()
        for component, optimization in optimizations.items():
            updated_mas[component] = await self._apply_optimization(mas_implementation[component], optimization)
        return updated_mas

    async def _apply_optimization(self, component: Dict[str, Any], optimization: Dict[str, Any]) -> Dict[str, Any]:
        # Implement logic to apply the optimization to the corresponding MAS component
        # This is a placeholder implementation
        optimized_component = component.copy()
        optimized_component["optimized"] = True
        optimized_component["optimization_applied"] = optimization["action"]
        return optimized_component

    async def provide_insights(self, performance_metrics: Dict[str, Any], optimizations: Dict[str, Any]) -> Dict[str, Any]:
        insights = {}
        for metric, value in performance_metrics.items():
            insight = await self._generate_insight(metric, value, optimizations)
            insights[metric] = insight
        return insights

    async def _generate_insight(self, metric: str, value: Any, optimizations: Dict[str, Any]) -> str:
        # Implement logic to generate insights based on metrics and optimizations
        # This is a placeholder implementation
        return f"Metric {metric} has value {value}. Consider applying optimization: {optimizations.get(metric, 'No specific optimization')}"

    async def optimize_mas(self, mas_implementation: Dict[str, Any], feedback: Dict[str, Any]) -> Dict[str, Any]:
        """
        Optimize the MAS based on feedback and performance metrics.

        Args:
            mas_implementation (Dict[str, Any]): The current implementation of the MAS.
            feedback (Dict[str, Any]): Feedback and performance metrics for the MAS.

        Returns:
            Dict[str, Any]: The optimized MAS implementation.
        """
        self.historical_data.append(feedback)
        
        if len(self.historical_data) > 10:  # Adjust as needed
            X = np.array([list(data.values()) for data in self.historical_data[:-1]])
            y = np.array([list(data.values()) for data in self.historical_data[1:]])
            self.optimization_model.fit(X, y)
        
        bottlenecks = await self._identify_bottlenecks(feedback)
        optimizations = await self.suggest_optimizations(bottlenecks)
        updated_mas = await self.implement_optimizations(mas_implementation, optimizations)
        
        if len(self.historical_data) > 10:
            predicted_performance = self.optimization_model.predict([list(feedback.values())])
            print(f"Predicted next state performance: {predicted_performance}")

        return updated_mas

    async def _identify_bottlenecks(self, feedback: Dict[str, Any]) -> List[str]:
        # Implement logic to identify bottlenecks based on feedback
        # This is a placeholder implementation
        bottlenecks = []
        for metric, value in feedback.items():
            if isinstance(value, (int, float)) and value > 0.8:  # Assuming higher values indicate potential bottlenecks
                bottlenecks.append(f"High {metric}")
        return bottlenecks
